fix: github action to lint (#102)
* fix linter action

* remove slinky from workflow

+++ x/opchild/keeper/executor_change_test.go
@@ -16,8 +16,8 @@ func Test_RegisterExecutorChangePlan(t *testing.T) {
 	_, input := createTestInput(t, false)
 
 	// Arguments
-	l1ProposalID := uint64(rand.Uint64())
-	height := uint64(rand.Uint64())
+	l1ProposalID := rand.Uint64()
+	height := rand.Uint64()
 	nextValAddr := valAddrsStr[0]
 	nextExecutorAddr := []string{addrsStr[0], addrsStr[1]}
 	consensusPubKey := "l7aqGv+Zjbm0rallfqfqz+3iN31iOmgJCafWV5pGs6o="
